bindings: play uses entities/types
