+++ pkg/domain/entities/play.go
@@ -4,6 +4,7 @@ import (
 	"net"
 
 	"github.com/containers/image/v5/types"
+	entitiesTypes "github.com/containers/podman/v4/pkg/domain/entities/types"
 )
 
 // PlayKubeOptions controls playing kube YAML files.
@@ -80,42 +81,14 @@ type PlayKubeOptions struct {
 }
 
 // PlayKubePod represents a single pod and associated containers created by play kube
-type PlayKubePod struct {
-	// ID - ID of the pod created as a result of play kube.
-	ID string
-	// Containers - the IDs of the containers running in the created pod.
-	Containers []string
-	// InitContainers - the IDs of the init containers to be run in the created pod.
-	InitContainers []string
-	// Logs - non-fatal errors and log messages while processing.
-	Logs []string
-	// ContainerErrors - any errors that occurred while starting containers
-	// in the pod.
-	ContainerErrors []string
-}
+type PlayKubePod = entitiesTypes.PlayKubePod
 
 // PlayKubeVolume represents a single volume created by play kube.
-type PlayKubeVolume struct {
-	// Name - Name of the volume created by play kube.
-	Name string
-}
+type PlayKubeVolume entitiesTypes.PlayKubeVolume
 
 // PlayKubeReport contains the results of running play kube.
-type PlayKubeReport struct {
-	// Pods - pods created by play kube.
-	Pods []PlayKubePod
-	// Volumes - volumes created by play kube.
-	Volumes []PlayKubeVolume
-	PlayKubeTeardown
-	// Secrets - secrets created by play kube
-	Secrets []PlaySecret
-	// ServiceContainerID - ID of the service container if one is created
-	ServiceContainerID string
-	// If set, exit with the specified exit code.
-	ExitCode *int32
-}
-
-type KubePlayReport = PlayKubeReport
+type PlayKubeReport = entitiesTypes.PlayKubeReport
+type KubePlayReport = entitiesTypes.KubePlayReport
 
 // PlayKubeDownOptions are options for tearing down pods
 type PlayKubeDownOptions struct {
@@ -124,13 +97,6 @@ type PlayKubeDownOptions struct {
 }
 
 // PlayKubeDownReport contains the results of tearing down play kube
-type PlayKubeTeardown struct {
-	StopReport     []*PodStopReport
-	RmReport       []*PodRmReport
-	VolumeRmReport []*VolumeRmReport
-	SecretRmReport []*SecretRmReport
-}
+type PlayKubeTeardown = entitiesTypes.PlayKubeTeardown
 
-type PlaySecret struct {
-	CreateReport *SecretCreateReport
-}
+type PlaySecret = entitiesTypes.PlaySecret
